service/pyproject.toml
[tool.poetry]
name = "llm-writer-workshop"
version = "0.1.0"
description = ""
authors = ["Jason Robison <jrrobison@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.3"
google-generativeai = "^0.4.1"
mistralai = "^0.1.6"
flask = "^3.0.2"
toml = "^0.10.2"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.1"
flask-injector = "^0.15.0"
anthropic = "^0.23.1"
marshmallow = "^3.21.1"


[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
pytest-flask = "^1.3.0"
pytest-mock = "^3.14.0"
pytest-cov = "5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
